refactor(core): Analyze, centralize, and document all core bot systems
This commit delivers a comprehensive analysis and improvement of the bot's core systems, including the main behavior loop (Attack, Target, Move, Loot) and primary support systems (Protection, Buff).

This is a consolidation of several iterative refactoring steps:

1.  **Centralized Attack Logic**: Extracted duplicated attack logic from multiple botbases into a new, central `AttackService` to eliminate code duplication.

2.  **Centralized Buff Logic**: Identified duplicated buffing logic and centralized it into a new `CastAvailableBuffs` method within the `SkillManager`, removing redundancy.

3.  **Analyzed and Documented Other Systems**:
    *   Conducted a thorough analysis of the `Target`, `Movement`, `Loot`, and `Protection` systems.
    *   Determined that these systems were either already well-centralized (`Loot`, `Protection`) or contained distinct, strategy-specific logic (`Target`, `Movement`) that should not be merged.
    *   As refactoring was not appropriate, detailed Turkish comments were added to all complex code sections within these modules to improve clarity and future maintainability, as requested by the user.
