+++ Plugins/RSBot.CommandCenter/Components/Command/BuffCommandExecutor.cs
@@ -1,7 +1,6 @@
-﻿using RSBot.Core;
+using RSBot.Core;
 using RSBot.Core.Components;
 using RSBot.Core.Components.Command;
-using RSBot.Core.Objects;
 
 namespace RSBot.CommandCenter.Components.Command;
 
@@ -16,22 +15,8 @@ internal class BuffCommandExecutor : ICommandExecutor
         if (!silent)
             Game.ShowNotification("[RSBot] Casting all buffs");
 
-        var buffs = SkillManager.Buffs.FindAll(p => !Game.Player.State.HasActiveBuff(p, out _) && p.CanBeCasted);
-        if (buffs.Count == 0)
-            return true;
-
-        Log.Status("Buffing");
-
-        foreach (var buff in buffs)
-        {
-            if (Game.Player.State.LifeState != LifeState.Alive || Game.Player.HasActiveVehicle ||
-                Game.Player.Untouchable)
-                break;
-
-            Log.Debug($"[ActionMapper] Casting buff {buff} ({buff.Record.Basic_Code})");
-
-            buff.Cast(buff: true);
-        }
+        // Merkezi buff basma metodunu çağır.
+        SkillManager.CastAvailableBuffs();
 
         return true;
     }
